requirements.txt
click>=4
docker-py>=1.4
requests==2.7.0
six==1.9.0
websocket-client==0.32.0
tabulate==0.7.5
